requirements.txt
eztcolors==0.3
moderngl==5.10.0
numpy==2.1.2
Pillow==11.0.0
pygame==2.5.2
PyGLM==2.7.1
pyrr==0.10.3
PyWavefront==1.3.3
tqdm==4.66.5